backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic-settings
bcrypt
email-validator
cryptography
